Добавил компоненты FileManagement, UserInfo, UserManagement. Добавил страницу UserFilesPage для отображения для админа списка файлов конкретного пользователя по ID. Добавил Возможность обновлять данные пользователя и изменять пароль. При логине если пользователь админ, то он будет перенаправлен на /admin и у него появится кнопка для управления пользователями и файлами. Обычный пользователь переходит в /dashboard

+++ backend/mycloud/mycloud/views.py
@@ -1,10 +1,11 @@
-from django.core.exceptions import ObjectDoesNotExist
+from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.response import Response
 from rest_framework.decorators import action, api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.views import APIView
 from rest_framework import status, viewsets
+from rest_framework.authtoken.models import Token
 from django.utils.timezone import now
 from django.conf import settings
 from django.http import FileResponse, Http404
@@ -14,6 +15,9 @@ from .models import File, TemporaryLink, CustomUser
 from .serializers import FileSerializer, UserSerializer
 import secrets
 import os
+from .validators import PasswordComplexityValidator
+from urllib.parse import quote
+
 
 
 class FileViewSet(viewsets.ModelViewSet):
@@ -23,7 +27,16 @@ class FileViewSet(viewsets.ModelViewSet):
 
     def get_queryset(self):
         """Фильтруем файлы по пользователю."""
-        return File.objects.filter(user=self.request.user)
+        user = self.request.user
+        user_id = self.request.query_params.get('user_id', None)
+
+        if user.is_staff and user_id:
+            return File.objects.filter(user_id=user_id)
+
+        if user.is_staff:
+            return File.objects.all()
+
+        return File.objects.filter(user=user)
 
     def perform_create(self, serializer):
         """Передаем текущего пользователя в сериализатор."""
@@ -33,11 +46,10 @@ class FileViewSet(viewsets.ModelViewSet):
     def download(self, request, pk=None):
         """Функция для скачивания файла"""
         file = self.get_object()
-        if file.user != request.user:
+        if file.user != request.user and not request.user.is_staff:
             return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
         file_path = os.path.join(settings.MEDIA_ROOT, file.file.name)
-
         # Проверяем, существует ли файл на сервере
         if not os.path.exists(file_path):
             return Response({"detail": "File not found"}, status=status.HTTP_404_NOT_FOUND)
@@ -53,18 +65,14 @@ class FileViewSet(viewsets.ModelViewSet):
         """Удаление файла"""
         try:
             file = self.get_object()
-
-
-            if file.user != request.user:
+            if file.user != request.user and not request.user.is_staff:
                 return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
             file_path = file.file.path
-
             if os.path.exists(file_path):
                 os.remove(file_path)
 
             file.delete()
-
             return Response({"detail": "File deleted successfully"}, status=status.HTTP_204_NO_CONTENT)
 
         except ObjectDoesNotExist:
@@ -74,12 +82,13 @@ class FileViewSet(viewsets.ModelViewSet):
     def rename_file(self, request, pk=None):
         """Переименование файла"""
         file = self.get_object()
-        if file.user != request.user:
+        if file.user != request.user and not request.user.is_staff:
             return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
         new_name = request.data.get("name")
         if not new_name:
             return Response({"detail": "Name field is required"}, status=status.HTTP_400_BAD_REQUEST)
+
         file.name = new_name
         file.save()
         return Response({"detail": "File renamed successfully", "new_name": file.name}, status=status.HTTP_200_OK)
@@ -88,7 +97,7 @@ class FileViewSet(viewsets.ModelViewSet):
     def update_comment(self, request, pk=None):
         """Обновление комментария к файлу"""
         file = self.get_object()
-        if file.user != request.user:
+        if file.user != request.user and not request.user.is_staff:
             return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
         comment = request.data.get('comment')
@@ -105,14 +114,13 @@ class FileViewSet(viewsets.ModelViewSet):
     def generate_link(self, request, pk=None):
         """Генерация временной ссылки на файл"""
         file = self.get_object()
-        if file.user != request.user:
+        if file.user != request.user and not request.user.is_staff:
             return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
         token = secrets.token_urlsafe(32)
-        expires_at = now() + timedelta(hours=1)  # Ссылка будет действительна 1 час
+        expires_at = now() + timedelta(hours=1) # Ссылка будет действительна 1 час
 
         temporary_link = TemporaryLink.objects.create(file=file, token=token, expires_at=expires_at)
-
         link = f"{request.build_absolute_uri('/')[:-1]}/api/files/temp/{token}/"
         return Response({
             "link": link,
@@ -135,9 +143,13 @@ class TemporaryLinkDownloadView(View):
         if not os.path.exists(file_path):
             raise Http404("File not found")
 
+        # Для поддержки русских символов
+        filename = quote(temporary_link.file.name)
+
         response = FileResponse(open(file_path, 'rb'))
         response['Content-Type'] = 'application/octet-stream'
-        response['Content-Disposition'] = f'attachment; filename="{temporary_link.file.name}"'
+        response['Content-Disposition'] = f'attachment; filename="{filename}"; filename*=utf-8\'\'{filename}'
+
         return response
 
 
@@ -165,22 +177,26 @@ def list_users(request):
     return Response(serializer.data, status=status.HTTP_200_OK)
 
 
-@api_view(['DELETE'])
+@api_view(['GET', 'DELETE'])
 @permission_classes([IsAuthenticated])
-def delete_user(request, pk):
-    """Удаление пользователя для администратора"""
-    if not request.user.is_superuser:
-        return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
-
+def user_detail(request, user_id):
+    """Получение или удаление пользователя по ID"""
     try:
-        user = CustomUser.objects.get(pk=pk)
-        user.delete()
-        return Response({"detail": "User deleted successfully"}, status=status.HTTP_204_NO_CONTENT)
+        user = CustomUser.objects.get(id=user_id)
     except CustomUser.DoesNotExist:
         return Response({"detail": "User not found"}, status=status.HTTP_404_NOT_FOUND)
 
+    if request.method == 'GET':
+        if not (request.user.is_staff or request.user.is_superuser):
+            return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
+        serializer = UserSerializer(user)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
-from rest_framework.authtoken.models import Token
+    if request.method == 'DELETE':
+        if not (request.user.is_staff or request.user.is_superuser):
+            return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
+        user.delete()
+        return Response({"detail": "User deleted successfully"}, status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(['POST'])
@@ -194,19 +210,13 @@ def logout(request):
     return Response({"detail": "Logged out successfully"}, status=status.HTTP_200_OK)
 
 
-class UserProfileView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request):
-        """Получение атрибутов пользователя"""
-        user = request.user
-        data = {
-            "username": user.username,
-            "is_admin": user.is_staff,
-            "email": user.email
-        }
-        return Response(data)
-
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def current_user_profile(request):
+    """Получение атрибутов текущего пользователя"""
+    user = request.user
+    serializer = UserSerializer(user)
+    return Response(serializer.data, status=status.HTTP_200_OK)
 
 class TokenVerifyView(APIView):
     authentication_classes = [TokenAuthentication]
@@ -251,3 +261,43 @@ def summary(self, request):
     }
 
     return Response(summary_data, status=status.HTTP_200_OK)
+
+
+@api_view(['PATCH'])
+@permission_classes([IsAuthenticated])
+def update_user_profile(request):
+    """Редактирование данных профиля пользователя"""
+    user = request.user
+    serializer = UserSerializer(user, data=request.data, partial=True)
+    
+    if serializer.is_valid():
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(['POST'])
+@permission_classes([IsAuthenticated])
+def change_password(request):
+    """Изменение пароля пользователя без старого пароля"""
+    user = request.user
+    new_password = request.data.get('new_password')
+    confirm_password = request.data.get('confirm_password')
+
+    # Проверка на совпадение пароля и его подтверждения
+    if new_password != confirm_password:
+        return Response({"detail": "Пароли не совпадают"}, status=status.HTTP_400_BAD_REQUEST)
+
+    # Проверка сложности пароля
+    password_validator = PasswordComplexityValidator()
+    try:
+        password_validator.validate(new_password)
+    except ValidationError as e:
+        return Response({"detail": str(e)}, status=status.HTTP_400_BAD_REQUEST)
+
+    # Установка нового пароля
+    user.set_password(new_password)
+    user.save()
+
+    return Response({"detail": "Пароль успешно изменён"}, status=status.HTTP_200_OK)
+
